feat: enhance customer meal suggestions and selection tips
- Updated the customer card and result card
to utilize the new unsatisfied selection tip for better user feedback.
- Integrated bond rewards into the info button for improved customer engagement.
- Refactored suggested meal card to leverage the new useSuggestedMealsViewModel hook
for better state management and performance.
- Added unsatisfied selection tip and bond rewards to the customerRareStore
for dynamic updates based on user selections.
- Improved documentation to reflect changes
in the current state and next steps for the project.
