+++ .github/plans/逻辑解耦/00-总体任务与进度.md
@@ -6,7 +6,7 @@
 - 当前阶段：阶段 0 / 阶段 1 / 阶段 2 / 阶段 3 / 阶段 4 / 阶段 5 / 阶段 6 文档已齐备
 - 当前焦点：阶段 1
 - 当前优先级：阶段 1
-- 当前里程碑：PR-1.2 已完成，下一步推进 PR-1.3
+- 当前里程碑：PR-1.3 已完成，下一步推进 PR-1.4
 - 最后更新：2026-04-23
 - 当前基线验证：`pnpm exec tsc --noEmit` 通过；`pnpm lint` 0 error / 12 warnings；`pnpm build` 通过，静态页 `137/137`
 
@@ -48,6 +48,6 @@
 ## 当前步骤
 
 1. 阶段 0 已完成：`app/utils/customer/shared/` 已落地 `PR-0.1 ~ PR-0.6` 的纯函数与 shared barrel，且保持旧页面/store 调用点不变。
-2. 阶段 1 的 PR-1.1 与 PR-1.2 已完成：`customer-rare` / `customer-normal` store 已承接表格 rows、`ingredientScoreChanges` 与 `savedCustomerMealsWithEvaluation`，稀客 resultCard 已改读 `currentMealPrice`。
-3. 下一步推进 PR-1.3：围绕 `unsatisfiedSelectionTip`、`bondRewards` 与 `useSuggestedMealsViewModel` 建立下一批承接层。
+2. 阶段 1 的 PR-1.1、PR-1.2、PR-1.3 已完成：`customer-rare` / `customer-normal` store 已承接表格 rows、`ingredientScoreChanges` 与 `savedCustomerMealsWithEvaluation`，稀客侧已补齐 `currentMealPrice`、`unsatisfiedSelectionTip`、`bondRewards` 与 `useSuggestedMealsViewModel`。
+3. 下一步推进 PR-1.4：围绕 `useCustomerRouteData` 与 `useIngredientRouteData` 收口 route 级业务 hook。
 4. 阶段 2 仍按 tab content -> saved meal -> rare cards/info -> route content 的顺序削薄视图，避免提前改动容器层。
