+++ app/hooks/index.ts
@@ -11,4 +11,5 @@ export { useOpenedItemPopover } from './useOpenedItemPopover';
 export * from './useSearchResult';
 export * from './useSkipProcessItemData';
 export * from './useSortedData';
+export * from './useSuggestedMealsViewModel';
 export { useViewInNewWindow } from './useViewInNewWindow';
